+++ __tests__/wrap-page.js
@@ -1,5 +1,4 @@
 const { preferDefault } = require(`../src/wrap-page`)
-const { polyfillIntl } = require(`../src/polyfill`)
 
 describe("preferDefault()", () => {
   describe("when module does not have default export", () => {
@@ -21,26 +20,6 @@ describe("preferDefault()", () => {
       const value = preferDefault(module)
 
       expect(value).toEqual("bar")
-    })
-  })
-})
-
-describe("polyfillIntl", () => {
-  const originalIntl = Intl
-
-  describe("when Intl is undefined", () => {
-    it("throws error", () => {
-      expect(() => {
-        polyfillIntl("not-a-language")
-      }).toThrow(`Cannot find react-intl/locale-data/not-a-language`)
-    })
-
-    beforeAll(() => {
-      Intl = undefined
-    })
-
-    afterAll(() => {
-      Intl = originalIntl
     })
   })
 })
